+++ apps/database-new/components/Header/SaveSchemaDropdown.tsx
@@ -1,0 +1,76 @@
+import { getAppStateSnapshot, useAppStateSnapshot } from '@/lib/state'
+import Image from 'next/image'
+import {
+  Button,
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuItem,
+  DropdownMenuTrigger,
+  IconChevronDown,
+  IconClipboard,
+  IconDownload,
+} from 'ui'
+
+const SaveSchemaDropdown = () => {
+  const snap = useAppStateSnapshot()
+
+  const copyToClipboard = () => {
+    const snap = getAppStateSnapshot()
+    const focused = window.document.hasFocus()
+    if (focused) {
+      window.navigator?.clipboard?.writeText(snap.selectedCode)
+    } else {
+      console.warn('Unable to copy to clipboard')
+    }
+  }
+
+  const downloadSQL = () => {
+    const snap = getAppStateSnapshot()
+    const blob = new Blob([snap.selectedCode], { type: 'text/plain' })
+    const url = URL.createObjectURL(blob)
+    const link = document.createElement('a')
+    link.download = 'migration.sql'
+    link.href = url
+    link.click()
+  }
+
+  const loadSQLInSupabase = () => {
+    const snap = getAppStateSnapshot()
+    window.open(
+      `https://supabase.com/dashboard/project/_/sql?content=${encodeURIComponent(
+        snap.selectedCode
+      )}`,
+      '_blank'
+    )
+  }
+
+  return (
+    <DropdownMenu>
+      <DropdownMenuTrigger asChild>
+        <Button
+          type="default"
+          onClick={() => snap.setHideCode(!snap.hideCode)}
+          iconRight={<IconChevronDown />}
+        >
+          Save schema
+        </Button>
+      </DropdownMenuTrigger>
+      <DropdownMenuContent className="w-44" align="center">
+        <DropdownMenuItem className="space-x-2" onClick={() => copyToClipboard()}>
+          <IconClipboard size={14} strokeWidth={2} />
+          <p>Copy SQL</p>
+        </DropdownMenuItem>
+        <DropdownMenuItem className="space-x-2" onClick={() => downloadSQL()}>
+          <IconDownload size={14} strokeWidth={2} />
+          <p>Download SQL</p>
+        </DropdownMenuItem>
+        <DropdownMenuItem className="space-x-2" onClick={() => loadSQLInSupabase()}>
+          <Image alt="supabase" src="/supabase.png" width={14} height={14} />
+          <p>Load SQL in Supabase</p>
+        </DropdownMenuItem>
+      </DropdownMenuContent>
+    </DropdownMenu>
+  )
+}
+
+export default SaveSchemaDropdown
